+++ trabalho-pratico/src/queries/q04.c
@@ -50,7 +50,7 @@ void *__q04_parse_arguments(size_t argc, char *const argv[argc]) {
  * @return A deep copy of @p args_data.
  */
 void *__q04_clone_arguments(const void *args_data) {
-    return (void *) args_data; /* Const cast acceptable - it's an just an integer */
+    return (void *) (size_t) args_data; /* Const cast acceptable - it's an just an integer */
 }
 
 /**
